+++ src/App.js
@@ -1,32 +1,7 @@
 import React from "react";
+import TodoList from "./TodoList";
+import AddTodoForm from "./AddTodoForm";
 
-/* define object */
-
-let todoList = [
-  {
-    id: 1,
-    title: "Complete assignment",
-  },
-  {
-    id: 2,
-    title: "Grocery",
-  },
-  {
-    id: 3,
-    title: "Walking",
-  },
-];
-
-/* define function TodoList */
-
-function funcTodoList(todoList) {
-  let arrayOutput = todoList.map(function (obj) {
-    // return <li key={obj.id}>{obj.id} - {obj.title}</li>
-    return <li key={obj.id}>{obj.title}</li>;
-  });
-
-  return arrayOutput;
-}
 
 /// ------------------- ///
 
@@ -35,11 +10,10 @@ function App() {
     <div>
       <h1>Todo List</h1>
 
-      <ul>
-        {todoList.map(function (obj) {
-          return <li key={obj.id}>{obj.title}</li>;
-        })}
-      </ul>
+      <TodoList />
+      <AddTodoForm />
+
+
     </div>
   );
 }
